refactor(build): 通用化构建期环境变量注入机制（issue #261）
- hvigor/subhub-secret-plugin.ts → build-env-inject-plugin.ts，遍历已声明的
  buildProfileFields 字段注入，扩展新变量无需改插件
- 注入键名 app.env.<FIELD> / APP_ENV_<FIELD>，优先级 local.properties → 环境变量 → 空默认值
- 同步更新 hvigorfile.ts、local.properties(.example)、release-build.yml、multiEnvBuild 文档

openspec: build-env-injection
